Fix an IF statement in the CMakeLists.txt.
The IF statement did not evaluate properly. It's checking for a variable
that holds the name for the library to be created by the CMake file. It
incorrectly evaluated the variable as being empty when in fact it was
not. The variable is set (does not have to) in one of the parent
CMakeLists.txt files.